// lumi.server/src/useCases/bill/dash/bill.dash.dto.ts
export type GetDashParams = {
  customer?: string;
  startDate?: Date;
  endDate?: Date;
  startEconomy?: number;
  endEconomy?: number;
  startEnergyConsumption?: number;
  endEnergyConsumption?: number;
  startEnergyCost?: number;
  endEnergyCost?: number;
  installationId?: number;
  customerId?: number;
};